SIG-1534 split DRF fields in V1 API app
