+++ api/app/signals/apps/api/app_settings.py
@@ -1,0 +1,1 @@
+SIGNALS_API_MAX_UPLOAD_SIZE = 8388608  # 8MB = 8*1024*1024
