+++ api/app/signals/apps/api/v1/serializers/nested/__init__.py
@@ -1,3 +1,7 @@
+"""
+API V1 nested serializers used in various 'signals.Signal' serializers.
+"""
+
 from signals.apps.api.v1.serializers.nested.attachment import _NestedAttachmentModelSerializer
 from signals.apps.api.v1.serializers.nested.category import _NestedCategoryModelSerializer
 from signals.apps.api.v1.serializers.nested.location import _NestedLocationModelSerializer
