+++ api/app/signals/apps/api/v1/serializers/signal_split.py
@@ -1,3 +1,6 @@
+"""
+Serializer splits `api.Signal` instances in several children.
+"""
 import copy
 from collections import OrderedDict
