+++ api/app/signals/apps/api/v1/views/signal.py
@@ -1,4 +1,3 @@
-
 from datapunt_api.rest import DatapuntViewSet, HALPagination
 from django_filters.rest_framework import DjangoFilterBackend
 from rest_framework.decorators import action
